requirements.txt
tensorflow
flask
streamlit
pillow
numpy
requests